Move messages tasks out of folder
