+++ augur/tasks/github/__init__.py
@@ -1,7 +1,7 @@
 from augur.tasks.github.contributors import *
 from augur.tasks.github.events import *
 from augur.tasks.github.issues import *
-from augur.tasks.github.messages.tasks import *
+from augur.tasks.github.messages import *
 from augur.tasks.github.pull_requests.tasks import *
 from augur.tasks.github.repo_info.tasks import *
 from augur.tasks.github.releases.tasks import *
