+++ augur/tasks/init/celery_app.py
@@ -30,7 +30,7 @@ github_tasks = ['augur.tasks.github.contributors',
                 'augur.tasks.github.issues',
                 'augur.tasks.github.pull_requests.tasks',
                 'augur.tasks.github.events',
-                'augur.tasks.github.messages.tasks',
+                'augur.tasks.github.messages',
                 'augur.tasks.github.facade_github.tasks',
                 'augur.tasks.github.releases.tasks',
                 'augur.tasks.github.repo_info.tasks',
